Move caching logic to a cacheExchange
